% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnames_shrink_ysn.R
\name{bnames_shrink_ysn}
\alias{bnames_shrink_ysn}
\title{Shrink Babynames
Filters Babynames dataframe to year/sex/name columns only}
\usage{
bnames_shrink_ysn()
}
\value{
dataframe
}
\description{
Shrink Babynames
Filters Babynames dataframe to year/sex/name columns only
}
\examples{
bnames_shrink_ysn()
}
