% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnames_shrink_ysn.R
\name{new.names}
\alias{new.names}
\title{New Name:
Returns three random baby names that are from a given year between 1880-2021
and also start with the first letter of a user-inputted name.}
\usage{
new.names()
}
\value{
list
}
\description{
New Name:
Returns three random baby names that are from a given year between 1880-2021
and also start with the first letter of a user-inputted name.
}
\examples{
new.name()
}
