% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{unisex}
\alias{unisex}
\title{Unisex Names Data}
\format{
\subsection{\code{unisex}}{

A data frame with 919 rows and 6 columns:
\describe{
\item{x1}{index}
\item{name}{first names from the Social Security Administration}
\item{total}{total number of people with that name}
\item{male_share}{percentage of males with that name}
\item{female_share}{percentage of females with that name}
\item{gap}{difference in percentage of males and females with a certain name}
...
}
}
}
\source{
\url{https://data.world/fivethirtyeight/unisex-names}
}
\usage{
unisex
}
\description{
Data that contains 900 names given to each gender at least
one third of the time with a minimum of 100 people with that name.
}
\keyword{datasets}
